+++ src/sound/mixer.h
@@ -90,16 +90,16 @@ const char *mixer_get_name(const int channel);
 */
 void mixer_set_volume(const int channel, const int volume);
 
-void mixer_play_sample(const int channel, INT8 *data, const int len, const int freq, const UINT8 loop);
-void mixer_play_sample_16(const int channel, INT16 *data, const int len, const int freq, const UINT8 loop);
+void mixer_play_sample(const int channel, INT8 *data, const int len, const double freq, const UINT8 loop);
+void mixer_play_sample_16(const int channel, INT16 *data, const int len, const double freq, const UINT8 loop);
 void mixer_stop_sample(const int channel);
 int mixer_is_sample_playing(const int channel);
-void mixer_set_sample_frequency(const int channel, const int freq);
+void mixer_set_sample_frequency(const int channel, const double freq);
 void mixer_sound_enable_global_w(const UINT8 enable);
 
-void mixer_play_streamed_sample_16(const int channel, INT16 *data, int len, const int freq);
+void mixer_play_streamed_sample_16(const int channel, INT16 *data, int len, const double freq);
 int mixer_samples_this_frame(void);
-int mixer_need_samples_this_frame(const int channel, const int freq);
+int mixer_need_samples_this_frame(const int channel, const double freq);
 
 /* private functions for user interface only - don't call them from drivers! */
 void mixer_set_mixing_level(const int channel,const int level);
